ai_server/requirement.txt
openai
pydantic
fastapi
uvicorn
python-dotenv